app/web/requirements.txt
boto3==1.38.27
fastapi==0.115.12
python-dotenv==1.1.0
redis==5.2.0
uvicorn==0.32.0